6.2.4.p2-0131-CL160660-fix-build-error-on-bg4ct
CL160660 fix build error on bg4ct

+++ driver/driver/khronos/libVulkan10/src/gc_vk_cmdbuf.c
@@ -712,7 +712,7 @@ VkResult __vk_CommitStateBuffers(
     for (icommitPool = 0; icommitPool <= curPoolIndex; ++icommitPool)
     {
         VkBool32 bLast = (icommitPool == curPoolIndex);
-        uint32_t count = bLast ? commitCount : (__VK_MAX_COMMITS << icommitPool);
+        uint32_t count = bLast ? commitCount : (uint32_t)(__VK_MAX_COMMITS << icommitPool);
         for (icommits = 0; icommits < count; icommits++)
         {
             uint32_t *states;
